Move API schema implementation to its own module
